requirements.txt
fastapi==0.75.0
matplotlib==3.2.2
Pillow==9.0.1
starlette==0.19.0
timm==0.5.4
torch==1.2.0
torchvision==0.4.0
uvicorn==0.17.6
